refactor: rename and organize modules
