--- src/harlequin/tui/__init__.py
@@ -1,12 +1,0 @@
-from harlequin.tui.code_editor import CodeEditor as CodeEditor
-from harlequin.tui.results_viewer import ResultsViewer as ResultsViewer
-from harlequin.tui.schema_viewer import (
-    SCHEMAS as SCHEMAS,
-)
-from harlequin.tui.schema_viewer import (
-    TABLES as TABLES,
-)
-from harlequin.tui.schema_viewer import (
-    SchemaViewer as SchemaViewer,
-)
-from harlequin.tui.error_modal import ErrorModal as ErrorModal
