+++ src/harlequin/tui/components/__init__.py
@@ -1,0 +1,12 @@
+from harlequin.tui.components.code_editor import CodeEditor as CodeEditor
+from harlequin.tui.components.results_viewer import ResultsViewer as ResultsViewer
+from harlequin.tui.components.schema_viewer import (
+    SCHEMAS as SCHEMAS,
+)
+from harlequin.tui.components.schema_viewer import (
+    TABLES as TABLES,
+)
+from harlequin.tui.components.schema_viewer import (
+    SchemaViewer as SchemaViewer,
+)
+from harlequin.tui.components.error_modal import ErrorModal as ErrorModal
