+++ src/harlequin/tui/components/code_editor.py
