+++ src/harlequin/tui/components/error_modal.py
