+++ src/harlequin/tui/components/results_viewer.py
