+++ src/harlequin/tui/components/schema_viewer.py
